// schemaregistry/test/serde/test/cycle_pb.ts
// @generated by protoc-gen-es v2.0.0 with parameter "target=ts"
// @generated from file test/schemaregistry/serde/cycle.proto (package test, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv1";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file test/schemaregistry/serde/cycle.proto.
 */
export const file_test_schemaregistry_serde_cycle: GenFile = /*@__PURE__*/
  fileDesc("CiV0ZXN0L3NjaGVtYXJlZ2lzdHJ5L3NlcmRlL2N5Y2xlLnByb3RvEgR0ZXN0IjsKCkxpbmtlZExpc3QSDQoFdmFsdWUYASABKAUSHgoEbmV4dBgKIAEoCzIQLnRlc3QuTGlua2VkTGlzdEIJWgcuLi90ZXN0YgZwcm90bzM");

/**
 * @generated from message test.LinkedList
 */
export type LinkedList = Message<"test.LinkedList"> & {
  /**
   * @generated from field: int32 value = 1;
   */
  value: number;

  /**
   * @generated from field: test.LinkedList next = 10;
   */
  next?: LinkedList;
};

/**
 * Describes the message test.LinkedList.
 * Use `create(LinkedListSchema)` to create a new message.
 */
export const LinkedListSchema: GenMessage<LinkedList> = /*@__PURE__*/
  messageDesc(file_test_schemaregistry_serde_cycle, 0);